Added better organisation for image service , and fixed route issues with recommendations , - neeed to check the scrappers perhaps not all are working accordingly
