+++ backend/requirements.txt
@@ -1,9 +1,9 @@
 aiohttp==3.10.5
 beautifulsoup4==4.12.3
 Flask==3.0.3
+flask_cors==5.0.0
 joblib==1.4.2
 numpy==2.1.1
 pandas==2.2.3
-Requests==2.32.3
 scikit_learn==1.5.2
 scipy==1.14.1
